Add CustomDrawer and related widgets, update DashboardDesktopLayout
Added a new `DrawerItemsModel` class to represent items in the drawer. Created `CustomDrawer` widget to display the drawer with user information and navigation items. Introduced `CustomListViewItems` to handle the list of drawer items with active and inactive states. Updated `DashboardDesktopLayout` to include the `CustomDrawer` and `AllExpenses` widget. Also, made minor adjustments to `UserInfoListTile` for styling consistency.